Share process test utilities
