+++ waspc/tests/Job/Process/LongRunningTest.hs
@@ -1,19 +1,19 @@
 module Job.Process.LongRunningTest where
 
-import Control.Concurrent (Chan, newChan, readChan, threadDelay)
+import Control.Concurrent (Chan, newChan, readChan)
 import qualified Control.Concurrent.Async as Async
 import Control.Exception (finally)
 import Control.Monad (void, when)
 import Data.Maybe (isJust)
 import qualified Data.Text as T
 import Data.Time.Clock (diffUTCTime, getCurrentTime)
-import System.Directory (doesFileExist, getTemporaryDirectory, removeFile)
+import System.Directory (doesFileExist, removeFile)
 import System.Exit (ExitCode (..))
-import System.IO (hClose, openTempFile)
 import System.Info (os)
 import qualified System.Process as P
 import System.Timeout (timeout)
 import Test.Hspec (Spec, describe, expectationFailure, it, shouldBe, shouldReturn, shouldSatisfy)
+import Test.Process.Util (isPortAvailable, makeTempPath, waitUntil)
 import qualified Wasp.Job as J
 import qualified Wasp.Job.Process.LongRunning as LongRunning
 import Wasp.Util (secondsToMicroSeconds)
@@ -203,63 +203,8 @@ collectQueuedOutput chan = go []
         Just J.JobMessage {J._data = J.JobOutput output _} -> go (output : collected)
         Just _ -> go collected
 
-makeTempPath :: String -> IO FilePath
-makeTempPath nameTemplate = do
-  tempDir <- getTemporaryDirectory
-  (filePath, fileHandle) <- openTempFile tempDir nameTemplate
-  hClose fileHandle
-  removeFile filePath
-  return filePath
-
-isProcessAlive :: String -> IO Bool
-isProcessAlive pid = do
-  (exitCode, _, _) <-
-    P.readCreateProcessWithExitCode
-      (P.proc "node" ["-e", "try { process.kill(Number(process.argv[1]), 0); } catch { process.exit(1); }", trim pid])
-      ""
-  return $ exitCode == ExitSuccess
-
-killProcess :: String -> IO ()
-killProcess pid =
-  void $
-    P.readCreateProcessWithExitCode
-      (P.proc "node" ["-e", "try { process.kill(Number(process.argv[1]), 'SIGKILL'); } catch {}", trim pid])
-      ""
-
-isPortAvailable :: String -> IO Bool
-isPortAvailable port = do
-  (exitCode, _, _) <-
-    P.readCreateProcessWithExitCode
-      (P.proc "node" ["-e", portProbeScript, trim port])
-      ""
-  return $ exitCode == ExitSuccess
-  where
-    portProbeScript =
-      unlines
-        [ "const net = require('node:net');",
-          "const server = net.createServer();",
-          "server.once('error', () => process.exit(1));",
-          "server.listen(Number(process.argv[1]), '127.0.0.1', () => server.close(() => process.exit(0)));"
-        ]
-
 jsString :: String -> String
 jsString = show
-
-trim :: String -> String
-trim = unwords . words
-
-waitUntil :: String -> IO Bool -> IO ()
-waitUntil label condition = go (50 :: Int)
-  where
-    go remainingAttempts
-      | remainingAttempts <= 0 = expectationFailure $ "Timed out waiting for " <> label
-      | otherwise = do
-          result <- condition
-          if result
-            then return ()
-            else do
-              threadDelay 100000
-              go $ remainingAttempts - 1
 
 removeFileIfExists :: FilePath -> IO ()
 removeFileIfExists filePath = do
